requirements.txt
streamlit
fastapi
uvicorn
langchain
openai
faiss-cpu
crewai
python-dotenv
pandas
requests
